pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "rag-mvp"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.50.0",
    "chromadb>=0.6.0",
    "pypdf>=4.0.0",
    "typer>=0.12.0",
    "rich>=13.0.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
rag = "src.main:app"

[tool.hatch.build.targets.wheel]
packages = ["src"]
